Connect animal map

+++ game/core/src/ntnu/gruppe22/game/helpers/GameInfo.java
@@ -8,6 +8,6 @@ public class GameInfo {
 
     public static final int WIDTH = 630;
     public static final int HEIGHT = 480;
-    public static final int PPM = 100;
+    public static final float PPM = 100;
 
 }
